need to find a way to make an argument array and pass that

+++ lol_threadz.h
@@ -10,7 +10,7 @@ typedef struct {
     int max;            // max amount of loops?
     int cur;            // current iteration
     char *data;         // data for the function pointer
-    long tid;           // thread id
+    int tid;            // thread id
 } lol_arg;
 
 typedef struct {
